refactor: some improves
