--- src/controllers/GPT/types.ts
@@ -1,9 +1,0 @@
-import { Context, NarrowedContext } from 'telegraf';
-import { Message, Update } from 'telegraf/typings/core/types/typegram';
-
-export type TextMatchedContext = NarrowedContext<
-  Context & { match: RegExpExecArray },
-  Update.MessageUpdate<Message.TextMessage>
->;
-
-export type TranscriptContext = NarrowedContext<Context, Update.MessageUpdate<Message.VoiceMessage>>;
