+++ src/controllers/types.ts
@@ -1,4 +1,11 @@
 import { Context, NarrowedContext } from 'telegraf';
-import { Update } from 'telegraf/typings/core/types/typegram';
+import { Message, Update } from 'telegraf/typings/core/types/typegram';
 
 export type BaseContext = NarrowedContext<Context, Update.MessageUpdate>;
+
+export type TextMatchedContext = NarrowedContext<
+  Context & { match: RegExpExecArray },
+  Update.MessageUpdate<Message.TextMessage>
+>;
+
+export type TranscriptContext = NarrowedContext<Context, Update.MessageUpdate<Message.VoiceMessage>>;
